+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Version 3.0.0
-* Support for games from chessgames.com and chess-db.com.
+* Support for games from chessgames.com, chess-db.com and chesstempo.com.
 * Improved manifest to limit permissions as much as possible.
 * Icon now becomes enabled only when there is a valid game to analyse.
 * Improved colour of icon so that it is clear when it is available.