+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Version 3.0.1
+* Workaround for chess.com bug where the 'result' field in the pgn is not updated after a player times out.
+
 ## Version 3.0.0
 * Support for games from chessgames.com, chess-db.com and chesstempo.com.
 * Improved manifest to limit permissions as much as possible.